state: Move selected snapshots out of diff parser
The diff parser provides unified-diff parsing while selected-change state
uses snapshot files to compare index and working tree content during
review.

The parser still owns selected-file snapshot persistence. That makes
parsing import repository lookup, editor buffer IO, session paths, and
journal logging even though those operations belong to selected-change
state.

This commit begins the selected-snapshot boundary split by moving snapshot
persistence into the data package and updating the existing callers to use
the new state module.

Subsequent commits will move the snapshot tests and add an import-boundary
check for the parser.
