+++ src/git_stage_batch/commands/show.py
@@ -7,7 +7,6 @@ import sys
 
 from ..batch.display import annotate_with_batch_source
 from ..core.diff_parser import acquire_unified_diff, build_line_changes_from_patch_lines
-from ..core.diff_parser import write_snapshots_for_selected_file_path
 from ..core.hashing import (
     compute_binary_file_hash,
     compute_gitlink_change_hash,
@@ -46,6 +45,7 @@ from ..data.file_review.state import (
 )
 from ..data.line_state import convert_line_changes_to_serializable_dict, load_line_changes_from_state
 from ..data.session import require_session_started
+from ..data.selected_change.snapshots import write_snapshots_for_selected_file_path
 from ..exceptions import exit_with_error
 from ..i18n import _
 from ..output import (
